Fixed functions using Chicago salary data examples
City of Chicago updated the formatting and column names
for the Current Employee Names, Salaries, and Position Titles
data set (xzkq-xp2w). Namely, renamed the annual salary column, which
broke a number of examples.

For this fix, I've done a "find" within the functions/ folder and
updated examples referncing the "employee_annual_salary" column and
renamed it to the correct name.

I have not looked for other examples at this point.