Standardize on "SODA Producer API" vs. "Publisher API" (#813)
"Publisher API" is ambiguous, particularly because we refer elsewhere to a "Publishing API" (which is actually DSMAPI), e.g., https://socratapublishing.docs.apiary.io. The actual API that dev.socrata.com is referring to is the write-side of the SODA API (/resource/4x4), and we already refer to the read-side as the "SODA Consumer API" (https://dev.socrata.com/consumers/getting-started.html). So let's follow that paradigm and standardize on "Producer/Consumer" language for the SODA API.