Disclude '?' if no args are present
We had been including a question mark at the end of tryit URLs regardless of whether there were query arguments present. This wasn't causing issues, AFAIK: it just looked weird when we rendered the URL with a question mark at the end of it. So let's disclude it if no query arguments are present.