Exclude '?' if no args are present (#824)
We had been including a question mark at the end of tryit URLs regardless of whether there were query arguments present. This wasn't causing issues, AFAIK: it just looked weird when we rendered the URL with a question mark at the end of it. So let's exclude it if no query arguments are present.