Use get instead of getByResourceName (#825)
getByResourceName returns a view object that does not do special things related to the OBE/NBE migration. For example, it returns a "location" column as a "point" column plus four location "sub-columns", whereas /views/4x4 returns it properly as a "location" column without the sub-columns (for backend-y migration reasons related to backwards compatibility).